requirements.txt
Flask==0.12
Jinja2==2.9.4
MarkupSafe==0.23
Werkzeug==0.11.15
click==6.7
gunicorn==19.6.0
itsdangerous==0.24
wsgiref==0.1.2
Flask_Login==0.4.0
Flask_Migrate==2.0.3
Flask_SQLAlchemy==2.2
Flask_WTF==0.14.2
psycopg2==2.7
SQLAlchemy==1.1.6
WTForms==2.1
BeautifulSoup4==4.5.3
requests==2.13.0
Flask_Cors==3.0.2
Flask_RESTful==0.2.12
lxml==3.7.3
PyJWT==1.5.0